// src/common/rbac/role.enum.ts
export enum Role {
  SUPER_ADMIN = 'SUPER_ADMIN',
  TOURNAMENT_MANAGER = 'TOURNAMENT_MANAGER',
  GAME_MANAGER = 'GAME_MANAGER',
  STATION_ADMIN = 'STATION_ADMIN',
  MANAGER = 'MANAGER',
  SUPPORT = 'SUPPORT',
  CLIENT = 'CLIENT'
}

export const STAFF_ROLES: Role[] = [
  Role.SUPER_ADMIN,
  Role.TOURNAMENT_MANAGER,
  Role.GAME_MANAGER,
  Role.STATION_ADMIN,
  Role.MANAGER,
  Role.SUPPORT
];